Complete master data

+++ app/Http/Controllers/ProductController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
 
@@ -11,18 +12,27 @@ class ProductController extends Controller
 {
   public function index()
   {
+    $this->authorize('viewAll', Product::class);
+
     return Inertia::render('Master/Products/Index', [
+      'can' => [
+        'create' => Gate::allows('create', Product::class),
+      ],
       'products' => Product::all()
     ]);
   }
 
   public function create()
   {
+    $this->authorize('create', Product::class);
+
     return Inertia::render('Master/Products/Create');
   }
 
   public function store(Request $request)
   {
+    $this->authorize('create', Product::class);
+
     $validated = $request->validate([
       'name' => 'required|unique:products,name',
       'description' => 'nullable',
@@ -38,6 +48,8 @@ class ProductController extends Controller
 
   public function show($id)
   {
+    $this->authorize('view', Product::class);
+
     return Inertia::render('Master/Products/Create', [
       "product" => Product::where("id", $id)->first()
     ]);
@@ -45,6 +57,8 @@ class ProductController extends Controller
 
   public function update(Request $request, Product $product)
   {
+    $this->authorize('update', Product::class);
+
     $validated = $request->validate([
       'name' => 'required|unique:products,name,' . $product->id,
       'description' => 'nullable',
@@ -60,6 +74,8 @@ class ProductController extends Controller
 
   public function destroy($id)
   {
+    $this->authorize('delete', Product::class);
+
     $ids = explode(',', $id);
     Product::whereIn('id', $ids)->delete();
     return Redirect::route('master.products.index');
